New endpoints for portalTraffic, portalConversionRate and RollUpShell change
Added portalTraffic which is a copy of sensorTraffic. Everything needs to
go to portalTraffic instead of sensorTraffic. Iphone app needs to be
changed to go to portalTraffic. RollupShell has been modified to do rollups
for portal devices as well, currently rollups are being done only for
devices with a network id.